@@ -1,6 +1,6 @@
 module github.com/geniussportsgroup/gateway_cache
 
-go 1.18
+go 1.20
 
 require (
 	github.com/pierrec/lz4 v2.6.1+incompatible
@@ -11,6 +11,7 @@ require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/frankban/quicktest v1.14.3 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/objx v0.1.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
